requirements.txt
Flask
grpcio
grpcio-tools
protobuf
gunicorn
qrcode[pil]
sqlite3
